csc/requirements.txt
amqp==5.2.0
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
billiard==4.2.0
Brotli==1.1.0
celery==5.4.0
certifi==2024.2.2
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
crispy-bootstrap5==2024.2
cron-descriptor==1.4.3
cryptography==43.0.1
cssselect2==0.7.0
Django==5.0.6
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-ckeditor==6.7.1
django-crispy-forms==2.1
django-js-asset==2.2.0
django-timezone-field==7.0
et-xmlfile==1.1.0
flower==2.0.1
fonttools==4.53.1
geographiclib==2.0
html5lib==1.1
humanize==4.10.0
idna==3.7
kombu==5.4.0
lxml==5.3.0
mysql-connector-python==9.0.0
mysqlclient==2.2.4
numpy==2.1.1
openpyxl==3.1.5
oscrypto==1.3.0
pandas==2.2.3
pillow==10.3.0
prometheus_client==0.20.0
prompt_toolkit==3.0.47
pycparser==2.22
pydyf==0.11.0
pyHanko==0.25.1
pyhanko-certvalidator==0.26.3
pypdf==5.0.0
pyphen==0.16.0
pypng==0.20220715.0
python-bidi==0.6.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
qrcode==7.4.2
razorpay==1.4.2
redis==5.0.8
reportlab==4.2.2
requests==2.32.2
setuptools==75.1.0
six==1.16.0
sqlparse==0.5.0
svglib==1.5.1
tinycss2==1.3.0
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
uritools==4.0.3
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
xhtml2pdf==0.2.16
zopfli==0.2.3
